refactor(chat): improve error propagation; remove silent failures and duplicate SSE logic
